+++ neraca_umkm/models/__init__.py
@@ -4,3 +4,7 @@ from . import neraca_budget
 from . import neraca_aging
 from . import neraca_cashflow
 from . import neraca_pl
+from . import neraca_category
+from . import account_move_line
+from . import neraca_category_report
+from . import neraca_recurring